Delete LICENSE
This was not originally intended to be a public repository. The code I use from other repositories does not have licenses, so I don't believe I should use an MIT license. I didn't consider that before creating this repo with an MIT license. For now, this repo will not have a license.